pythonScript/requirements.txt
Flask==2.0.1
gunicorn==20.1.0
numpy==1.21.1
Pillow==8.2.0
requests==2.25.1
tensorflow-cpu==2.7.0